# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Microservice para gestão de cobranças (inicialmente integração Asaas; no futuro bancos diversos).",
        "title": "Charges Service",
        "contact": {},
        "version": "0.1.0"
    },
    "host": "localhost:8083",
    "basePath": "/",
    "paths": {
...
        "/v1/asaas/customers/by-company": {
            "get": {
                "description": "Recupera o customer do Asaas usando o mapeamento salvo em company.asaas_integration (via RPC em public), e então consulta o Asaas. Útil porque o front normalmente tem company_id. Referência Asaas: https://docs.asaas.com/reference/recuperar-um-unico-cliente",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "asaas"
                ],
                "summary": "Recuperar cliente no Asaas por company_id (mapeamento interno)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do accounting_office (UUID)",
                        "name": "accounting_office_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID da empresa (company_id, UUID)",
                        "name": "company_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.AsaasCustomerResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "description": "Resolve o customer id do Asaas via company.asaas_integration (RPC em public) e atualiza o cliente no Asaas. Referência Asaas: https://docs.asaas.com/reference/atualizar-cliente-existente",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "asaas"
                ],
                "summary": "Atualizar cliente existente no Asaas por company_id (mapeamento interno)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do accounting_office (UUID)",
                        "name": "accounting_office_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID da empresa (company_id, UUID)",
                        "name": "company_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Campos para atualização (parcial)",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.AsaasUpdateCustomerRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.AsaasCustomerResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/v1/asaas/customers/{id}": {
            "get": {
                "description": "Recupera um cliente no Asaas pelo ID (Asaas customer id), usando a integração ativa (is_active=true) do escritório (accounting_office_id) cadastrada no Supabase (iam.billing_integrations). Referência Asaas: https://docs.asaas.com/reference/recuperar-um-unico-cliente",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "asaas"
                ],
                "summary": "Recuperar um único cliente no Asaas",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do accounting_office (UUID)",
                        "name": "accounting_office_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID do cliente no Asaas",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.AsaasCustomerResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "description": "Atualiza um cliente existente no Asaas pelo ID (Asaas customer id), usando a integração ativa (is_active=true) do escritório (accounting_office_id) cadastrada no Supabase (iam.billing_integrations). Referência Asaas: https://docs.asaas.com/reference/atualizar-cliente-existente",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "asaas"
                ],
                "summary": "Atualizar cliente existente no Asaas",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do accounting_office (UUID)",
                        "name": "accounting_office_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID do cliente no Asaas",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Campos para atualização (parcial)",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.AsaasUpdateCustomerRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.AsaasCustomerResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "model.AsaasCustomerResponse": {
            "type": "object",
            "properties": {
                "additionalEmails": {
                    "type": "string"
                },
                "address": {
                    "type": "string"
                },
                "addressNumber": {
                    "type": "string"
                },
                "city": {
                    "type": "integer"
                },
                "cityName": {
                    "type": "string"
                },
                "complement": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "cpfCnpj": {
                    "type": "string"
                },
                "dateCreated": {
                    "type": "string"
                },
                "deleted": {
                    "type": "boolean"
                },
                "email": {
                    "type": "string"
                },
                "externalReference": {
                    "type": "string"
                },
                "foreignCustomer": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "mobilePhone": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "notificationDisabled": {
                    "type": "boolean"
                },
                "object": {
                    "type": "string"
                },
                "observations": {
                    "type": "string"
                },
                "personType": {
                    "description": "JURIDICA | FISICA",
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "postalCode": {
                    "type": "string"
                },
                "province": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                }
            }
        },
...
        "model.AsaasPersonType": {
            "type": "string",
            "enum": [
                "JURIDICA",
                "FISICA"
            ],
            "x-enum-varnames": [
                "AsaasPersonTypeJuridica",
                "AsaasPersonTypeFisica"
            ]
        },
        "model.AsaasUpdateCustomerRequest": {
            "type": "object",
            "properties": {
                "additionalEmails": {
                    "description": "Metadados",
                    "type": "string"
                },
                "address": {
                    "description": "Endereço",
                    "type": "string"
                },
                "addressNumber": {
                    "type": "string"
                },
                "city": {
                    "type": "integer"
                },
                "company": {
                    "type": "boolean"
                },
                "complement": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "cpfCnpj": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "externalReference": {
                    "type": "string"
                },
                "foreignCustomer": {
                    "type": "boolean"
                },
                "mobilePhone": {
                    "type": "string"
                },
                "name": {
                    "description": "Identificação",
                    "type": "string"
                },
                "notificationDisabled": {
                    "description": "Use pointers for booleans so \"omit vs false\" is preserved.",
                    "type": "boolean"
                },
                "observations": {
                    "type": "string"
                },
                "personType": {
                    "description": "Flags / enums",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.AsaasPersonType"
                        }
                    ]
                },
                "phone": {
                    "type": "string"
                },
                "postalCode": {
                    "type": "string"
                },
                "province": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                }
            }
        },
...
    }
}